requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1
django-cleanup==4.0.0
django-crispy-forms==1.7.2
idna==2.7
Pillow==7.0.0
pytz==2018.5